Installs bootstrap with typescript
